requirements.txt
pydantic
requests
aiohttp
beautifulsoup4
PyJWT
